Move ParticipantAvatars out of "internal" package
